reupload

+++ q2/utils.py
@@ -13,7 +13,7 @@ def load_train():
     with open('data/mnist_train.npz', 'rb') as f:
         train_set = np.load(f)
         train_inputs = train_set['train_inputs']
-        train_targets = train_set['train_targets']
+        train_targets = train_set['train_targets'].astype(np.float64)
 
     return train_inputs, train_targets
 
@@ -24,7 +24,7 @@ def load_train_small():
     with open('data/mnist_train_small.npz', 'rb') as f:
         train_set_small = np.load(f)
         train_inputs_small = train_set_small['train_inputs_small']
-        train_targets_small = train_set_small['train_targets_small']
+        train_targets_small = train_set_small['train_targets_small'].astype(np.float64)
 
     return train_inputs_small, train_targets_small
 
@@ -35,7 +35,7 @@ def load_valid():
     with open('data/mnist_valid.npz', 'rb') as f:
         valid_set = np.load(f)
         valid_inputs = valid_set['valid_inputs']
-        valid_targets = valid_set['valid_targets']
+        valid_targets = valid_set['valid_targets'].astype(np.float64)
 
     return valid_inputs, valid_targets
 
@@ -46,6 +46,6 @@ def load_test():
     with open('data/mnist_test.npz', 'rb') as f:
         test_set = np.load(f)
         test_inputs = test_set['test_inputs']
-        test_targets = test_set['test_targets']
+        test_targets = test_set['test_targets'].astype(np.float64)
 
     return test_inputs, test_targets
